FEAT: completed urlifyier! URLS can now be produce d from the converted json files!
FEAT: view-select-downloader received additional
functionality to accept and tackle any v3
manifests which have nested v2 image tags;
fallback is, within a v3 manifest, to use v2
images parsing if something goes wrong.

FEAT: added a default auto-detect functionality
to the VSD, with optional override

TO-DO: add a more thorough recommendation
for using github pages as a personal hosting
solution for the generated static site for users
who are using these manifests for personal
projects and research.
OPTIONALLY: add a question link for users to
visit more comprehensive documentation on how
to do this, if they're unfamiliar with github
and github pages.